% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcomp.R
\name{diffcomp}
\alias{diffcomp}
\title{Calculate differential compartments}
\usage{
diffcomp(data, colData, featureData, bioreps, samples, cutoff = 0.05)
}
\arguments{
\item{data}{data.frame with columns of compartment scores and rows for each
Hi-C bin.}

\item{colData}{data.frame describing the columns of \code{data}. It must
contain columns describing samples and bioreps. \code{rownames(colData)}
should equal \code{colnames(data)}.}

\item{featureData}{(optional) data.frame describing the rows of \code{data}.
Each column of \code{featureData} should describe different features of the
corresponding rows in \code{data}. \code{nrow(featureData)} should equal
\code{nrow(data)}.}

\item{bioreps}{string naming a column in \code{colData} that describes which
rows belong to which biological replicate.}

\item{samples}{string naming a column in \code{colData} that describes which
rows belong to which sample.}

\item{cutoff}{numeric threshold for false-discovery rate.}
}
\value{
Returns two data.frames. \code{distdf} contains \code{featureData},
  \code{data}, and the pair-wise distances between bioreps and samples along
  with a column denoting significance (T/F). \code{ranked} contains a ranked
  listing of samples, bioreps, fdr, and rank.
}
\description{
This function uses biological replicates (bioreps) to calculate a
false-discovery rate (fdr) for determining differential compartments between
two samples.
}
\details{
Pair-wise euclidean distances between input vectors are calculated between
sample-matched biological replicates (bioreps) and biorep-matched samples.
All distance vectors are then combined and ranked by decreasing value. The
false-discovery rate (fdr) is calculated cumulatively along the ranked
distance vector as a percentage of bioreps. Significance is determined by
finding the greatest rank that passes the fdr cutoff threshold.
}
\examples{
## Load example compartment data
data("compartments")

## Define eigenvector data
eigen <- compartments[,grep("EI", colnames(compartments))]

## Define colData describing samples and bioreps
cData <- data.frame(
  time = c("0000", "0000", "1440", "1440"),
  br = c(1, 2, 1, 2),
  row.names = colnames(eigen)
)

## Define featureData describing rows of data
fData <- compartments[,1:3]

## Calculate differential compartments
res <- diffcomp(data = eigen,
                colData = cData,
                featureData = fData,
                bioreps = "br",
                samples = "time",
                cutoff = 0.01)

## Look at results
lapply(res, head)

## Filter for significant compartments
sigComp <- res$distdf[res$distdf$significant == T,]
head(sigComp)
nrow(sigComp)



}
